index: Extract `credentials` module
